% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{agriland}
\alias{agriland}
\title{Agricultural Lands}
\format{A data.frame object with 254 rows and 3 columns
\describe{
  \item{\code{country_name}}{: name of the country.}
  \item{\code{country_code}}{: the iso3 code identifying the country.}
  \item{\code{percent}}{: percentage of lang.}
}}
\source{
\url{https://data.worldbank.org/indicator/AG.LND.AGRI.ZS}
}
\usage{
agriland
}
\description{
A datatset of the available agricultural land (as percentage of land area) by country in 2016.
}
\keyword{datasets}
