% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{animation}
\alias{animation}
\alias{pause_animation}
\alias{resume_animation}
\title{Point of View}
\usage{
pause_animation(globe)

resume_animation(globe)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}
}
\description{
Change the camera point of view.
}
