% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{bars_data}
\alias{bars_data}
\alias{bars_lat}
\alias{bars_lon}
\alias{bars_color}
\alias{bars_altitude}
\alias{bars_radius}
\alias{bars_resolution}
\alias{bars_merge}
\alias{bars_transition}
\alias{bars_on_click}
\alias{bars_on_right_click}
\alias{bars_on_hover}
\title{Points Functional API}
\usage{
bars_data(globe, data)

bars_lat(globe, lat = "lat")

bars_lon(globe, lon = "lng")

bars_color(globe, color = constant("ffffaa"))

bars_altitude(globe, altitude = 0.1)

bars_radius(globe, radius = 0.25)

bars_resolution(globe, resolution = 12L)

bars_merge(globe, merge = TRUE)

bars_transition(globe, transition = 1000L)

bars_on_click(globe, func)

bars_on_right_click(globe, func)

bars_on_hover(globe, func)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame of points to draw.}

\item{lat, lon}{Column names or numeric value indicating coordinates.}

\item{color}{Column name or character vector indicating color of points.}

\item{altitude}{Column name or character vector indicating altitude of points 
in terms of globe radius units (0 = 0 altitude (flat circle), 1 = globe radius).}

\item{radius}{Column name of radius a numeric constant for the cylinder's 
radius, in angular degrees.}

\item{resolution}{Numeric value defining the geometric resolution of each 
cylinder, expressed in how many slice segments to divide the circumference. 
Higher values yield smoother cylinders.}

\item{merge}{Whether to merge all the point meshes 
into a single ThreeJS object, for improved rendering performance. 
Visually both options are equivalent, setting this option only affects 
the internal organization of the ThreeJS objects.}

\item{transition}{Duration (ms) of the transition 
to animate point changes involving geometry modifications. A value of 
\code{0} will move the objects immediately to their final position. 
New objects are animated by scaling them from the ground up. Only works 
if \code{merge} is disabled.}

\item{func}{JavaScript function as character vector.}
}
\description{
Functional API to add and customise points on globe.
}
\examples{
# use data
create_globe() \%>\% 
  bars_data(quakes) \%>\% 
  bars_lat("lat") \%>\% 
  bars_lon("long")

# use in shiny
library(shiny)

ui <- fluidPage(
  actionButton("draw", "draw points"),
  globeOutput("globe")
)

server <- function(input, output) {
  output$globe <- renderGlobe({
    create_globe() \%>\% 
      bars_color(constant("#ffffff")) 
  })

  observeEvent(input$draw, {
    globeProxy("globe") \%>\% 
      bars_data(quakes) \%>\% 
      bars_lon("long") \%>\% 
      globe_pov(-21, 179)
  })
}

\dontrun{shinyApp(ui, server)}

}
