% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cables}
\alias{cables}
\title{Cables}
\format{An object of class sp containing LINESTRINGS.}
\source{
\url{https://raw.githubusercontent.com/telegeography/www.submarinecablemap.com/master/public/api/v2/cable/cable-geo.json}
}
\usage{
cables
}
\description{
Underwater cables.
}
\keyword{datasets}
