% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\title{Coordinates}
\usage{
coords(lat, lon, ...)
}
\arguments{
\item{lat, lon, ...}{List of name value pairs giving aspects to map to
variables. The names for lat and lon aspects are typically omitted because
they are so common; all other aspects must be named.}
}
\description{
Chart's aspects, similar to \code{ggplot2}'s aesthetics.
}
\section{Figures}{

\itemize{
  \item{\code{lat}, \code{lon}}
}
}

