% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe.R
\name{create_globe}
\alias{create_globe}
\title{Create a globe}
\usage{
create_globe(data = NULL, coords = NULL, antialias = TRUE,
  alpha = TRUE, animate_in = TRUE, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing coordinates (\code{\link{coords}}).}

\item{coords}{Coordinates as specified by \code{\link{coords}}.}

\item{antialias, alpha}{Configuration parameters to pass to the 
\href{https://threejs.org/docs/#api/en/renderers/WebGLRenderer}{ThreeJS WebGLRenderer} constructor.}

\item{animate_in}{Whether to animate the globe initialization, by scaling 
and rotating the globe into its inital position.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{Id of element.}
}
\description{
Create a globe object.
}
