% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globe.R
\name{globe-shiny}
\alias{globe-shiny}
\alias{globeOutput}
\alias{renderGlobe}
\alias{render_globe}
\alias{globeProxy}
\alias{globe_proxy}
\title{Shiny bindings for globe}
\usage{
globeOutput(outputId, width = "100\%", height = "400px")

renderGlobe(expr, env = parent.frame(), quoted = FALSE)

render_globe(expr, env = parent.frame(), quoted = FALSE)

globeProxy(id, session = shiny::getDefaultReactiveDomain())

globe_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a globe}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}

\item{id}{Target chart id.}

\item{session}{Shiny session.}
}
\description{
Output and render functions for using globe within Shiny
applications and interactive Rmd documents.
}
