% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{globe_choropleth}
\alias{globe_choropleth}
\title{Choropleth}
\usage{
globe_choropleth(
  globe,
  ...,
  data = NULL,
  inherit_coords = TRUE,
  on_click = NULL,
  on_right_click = NULL,
  on_hover = NULL,
  match = c("auto", "iso2", "iso3", "name"),
  polygons = NULL
)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{...}{Coordinates, as specified by \code{\link{coords}}.}

\item{data}{A data.frame of points to draw.}

\item{inherit_coords}{Whether to inherit the coordinates (\code{\link{coords}})
from \code{\link{create_globe}}. Only applies to method applied to object of class
\code{globe4r} not on objects of class \code{globeProxy}.}

\item{on_click}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{on_right_click}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{on_hover}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{match}{The type of \code{country} identifier, 
\code{auto} attempts to infer the type, \code{iso2}
(e.g.: "US"), iso3 (e.g.: "USA"), or the country name.}

\item{polygons}{Polygons to match data against, if
\code{NULL} is matched against internal dataset of 
world country polygons.}
}
\description{
Add points to a globe.
}
\details{
By default the function will match the 
\code{\link{coords}} \code{country} argument 
against an internal dataset of country names,
and iso codes.
In order to use other polygons, one can pass
GeoJSON as a \code{list} to the \code{polygons}
argument and specify \code{polygon} argument in
\code{\link{coords}}: this will match the latter
against the \code{data}, see
\href{https://globe4r.john-coene.com/articles/polygons.html}{online documentation}
for a full example.
}
\section{Coordinates}{

Valid coordinates.
\itemize{
  \item{\code{country}}
  \item{\code{polygon}}
  \item{\code{altitude}}
  \item{\code{label}}
  \item{\code{side_color}, \code{cap_color}}
  \item{\code{transition}}
}
}

\examples{
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_choropleth(
    data = agriland, 
    coords(country = country_code, cap_color = percent)
  )

}
