% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{globe_labels}
\alias{globe_labels}
\title{Points}
\usage{
globe_labels(globe, data, lat = NULL, lon = NULL, color = NULL,
  text = NULL, label = NULL, altitude = NULL, size = NULL,
  rotation = NULL, include_dot = NULL, dot_radius = NULL,
  dot_orientation = NULL, resolution = 2L, type_face = NULL,
  transition = 1000L, on_click = NULL, on_right_click = NULL,
  on_hover = NULL)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame of points to draw.}

\item{lat}{Bare column names of points coordinates.}

\item{lon}{Bare column names of points coordinates.}

\item{color}{Bare column name of points color.}

\item{text}{Bare column name containing text to display on globe.}

\item{label}{Bare column name of points label. 
Supports plain text or HTML content.}

\item{altitude}{Bare column name of points
defining the cylinder's altitude in terms of globe radius units 
(0 = 0 altitude (flat circle), 1 = globe radius).}

\item{size}{Bare column name of or the label text height, in angular degrees.}

\item{rotation}{Bare column name for the label rotation in degrees. 
The rotation is performed clockwise along the axis of its latitude parallel plane.}

\item{include_dot}{Bare column name for for whether to include a dot marker next 
to the text indicating the exact \code{lat}, \code{lon} coordinates of the label. 
If enabled the text will be rendered offset from the dot.}

\item{dot_radius}{Bare column name for the radius of the dot marker, in angular degrees.}

\item{dot_orientation}{Bare column name for the orientation of the label if the 
dot marker is present. Possible values are right, top and bottom.}

\item{resolution}{Getter/setter for the radial geometric resolution 
of each cylinder, expressed in how many slice segments to divide the
 circumference. Higher values yield smoother cylinders.}

\item{type_face}{Typeface to use for labels. Supports any typeface font 
generated by \href{http://gero3.github.io/facetype.js/}{Facetype.js}.}

\item{transition}{Getter/setter for duration (ms) of the transition 
to animate point changes involving geometry modifications. A value of 
\code{0} will move the objects immediately to their final position. 
New objects are animated by scaling them from the ground up. Only works 
if \code{merge} is disabled.}

\item{on_click}{JavaScript functions as strings.}

\item{on_right_click}{JavaScript functions as strings.}

\item{on_hover}{JavaScript functions as strings.}
}
\description{
Add points to a globe.
}
\examples{
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_labels(
    quakes,
    lat,
    long,
    text = stations
  )

library(shiny)

ui <- fluidPage(
  actionButton("add", "add quakes labels"),
  globeOutput("globe")
)

server <- function(input, output){
  output$globe <- renderGlobe({
    create_globe() \%>\% 
      globe_img_url() 
  })

  observeEvent(input$add, {
    globeProxy("globe") \%>\% 
      globe_labels(
        quakes,
        lat,
        long,
        text = stations
      )
  })
}

\dontrun{shinyApp(ui, server)}

}
