% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{globe_paths}
\alias{globe_paths}
\title{paths}
\usage{
globe_paths(
  globe,
  ...,
  data = NULL,
  inherit_coords = TRUE,
  on_click = NULL,
  on_right_click = NULL,
  on_hover = NULL
)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{...}{Coordinates, as specified by \code{\link{coords}}.}

\item{data}{A data.frame of points to draw.}

\item{inherit_coords}{Whether to inherit the coordinates (\code{\link{coords}})
from \code{\link{create_globe}}. Only applies to method applied to object of class
\code{globe4r} not on objects of class \code{globeProxy}.}

\item{on_click}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{on_right_click}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{on_hover}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}
}
\description{
Add paths to a globe.
}
\section{Coordinates}{

Valid coordinates.
\itemize{
  \item{\code{label}}
  \item{\code{altitude}}
  \item{\code{color}}
  \item{\code{resolution}}
  \item{\code{stroke}}
  \item{\code{dash_length}}
  \item{\code{initial_dash_length}}
  \item{\code{dash_animate_time}}
  \item{\code{transition}}
}
}

\examples{
create_globe(cables) \%>\% 
 globe_paths(
   coords(
     color = color, 
     dash_length = .1, 
     dash_gap = .008, 
     dash_animate_time = 12000
   )
 )

}
