% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{globe_points}
\alias{globe_points}
\title{Points}
\usage{
globe_points(globe, ..., data = NULL, inherit_coords = TRUE,
  on_click = NULL, on_right_click = NULL, on_hover = NULL)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{...}{Coordiantes, as specified by \code{\link{coords}}.}

\item{data}{A data.frame of points to draw.}

\item{inherit_coords}{Whether to inherit the coordinates (\code{\link{coords}})
from \code{\link{create_globe}}}

\item{on_click, on_right_click, on_hover}{JavaScript functions as strings.}
}
\description{
Add points to a globe.
}
\examples{
# basic
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_pov(-21, 179) \%>\% 
  globe_points(coords(lat, long, label = stations), data = quakes)

# use in shiny
library(shiny)

ui <- fluidPage(
  actionButton("add", "Add points"),
  globeOutput("globe")
)

server <- function(input, output){
  output$globe <- renderGlobe({
    create_globe() \%>\% globe_img_url()
  })

  observeEvent(input$add, {
    globeProxy("globe") \%>\% 
      globe_points(coords(lat, long), data = quakes) \%>\% 
      globe_pov(-21, 179)
  })
}

\dontrun{shinyApp(ui, server)}
}
