% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.R
\name{globe_pov}
\alias{globe_pov}
\title{Point of View}
\usage{
globe_pov(globe, lat, lon, altitude = 2.5)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{lat, lon}{Coordinates of camera position.}

\item{altitude}{Altitude of camera, defaults to \code{2.5} globe radii.}
}
\description{
Change the camera point of view.
}
\examples{
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_pov(38, -77)

}
