% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/controls.R
\name{globe_rotate}
\alias{globe_rotate}
\title{Controls}
\usage{
globe_rotate(globe, speed = 1L, rotate = TRUE)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{speed}{How fast to rotate around the globe.}

\item{rotate}{Whether to autorotate the globe.}
}
\description{
Controls globe.
}
\examples{
# basic
create_globe() \%>\% 
  globe_pov(-21, 179) \%>\% 
  globe_bars(
    coords(lat, long), 
    data = quakes
  ) \%>\% 
  globe_rotate()

}
