% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title.R
\name{globe_title}
\alias{globe_title}
\title{Title}
\usage{
globe_title(globe, text, tag = shiny::tags$h1, vertical = c("top",
  "bottom"), horizontal = c("left", "right"), color = "#fff",
  margin = "3\%")
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{text}{The text to display as title or subtitle.}

\item{tag}{The html tag to use, from \link[shiny]{tags}.}

\item{vertical, horizontal}{Alignment of title.}

\item{color}{Color of title or title.}

\item{margin}{Margin from \code{horizontal} and \code{vertical}.}
}
\description{
Add a title to a globe.
}
\examples{
create_globe() \%>\% 
  globe_pov(-21, 179) \%>\% 
  globe_bars(
    coords(lat, long, label = stations), 
    data = quakes
  ) \%>\% 
  globe_title("A Globe!")

}
