% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hex.R
\name{hex_data}
\alias{hex_data}
\alias{hex_lat}
\alias{hex_lon}
\alias{hex_weight}
\title{Hex Raw API}
\usage{
hex_data(globe, data)

hex_lat(globe, lat = "lat")

hex_lon(globe, lon = "lng")

hex_weight(globe, weight = 1L)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame of points to draw.}

\item{lat, lon}{Column names or numeric value indicating coordinates.}

\item{color}{Column name or character vector indicating color of points.}

\item{label}{Column name or constant of label.}

\item{altitude}{Column name or character vector indicating altitude of points 
in terms of globe radius units (0 = 0 altitude (flat circle), 1 = globe radius).}

\item{radius}{Column name of radius a numeric constant for the cylinder's 
radius, in angular degrees.}

\item{resolution}{Numeric value defining the geometric resolution of each 
cylinder, expressed in how many slice segments to divide the circumference. 
Higher values yield smoother cylinders.}

\item{merge}{Whether to merge all the point meshes 
into a single ThreeJS object, for improved rendering performance. 
Visually both options are equivalent, setting this option only affects 
the internal organization of the ThreeJS objects.}

\item{transition}{Duration (ms) of the transition 
to animate point changes involving geometry modifications. A value of 
\code{0} will move the objects immediately to their final position. 
New objects are animated by scaling them from the ground up. Only works 
if \code{merge} is disabled.}

\item{func}{JavaScript function as character vector.}
}
\description{
Functional API to add and customise hex on globe.
}
\examples{
# use data
create_globe() \%>\% 
  hex_data(quakes) \%>\% 
  hex_lat("lat") \%>\% 
  hex_lon("long") \%>\% 
  hex_weight("mag")

}
