% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels_data}
\alias{labels_data}
\alias{labels_lat}
\alias{labels_lon}
\alias{labels_text}
\alias{labels_color}
\title{Points Functional API}
\usage{
labels_data(globe, data)

labels_lat(globe, lat = "lat")

labels_lon(globe, lon = "lng")

labels_text(globe, text = "text")

labels_color(globe, color = constant("#ffffaa"))
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame of points to draw.}

\item{lat}{Column names or numeric value indicating coordinates.}

\item{lon}{Column names or numeric value indicating coordinates.}

\item{text}{Column name or constant of text.}

\item{color}{Column name or character vector indicating color of points.}
}
\description{
Functional API to add and customise points on globe.
}
\examples{
# use data
create_globe() \%>\% 
  labels_data(quakes) \%>\% 
  labels_lat("lat") \%>\% 
  labels_lon("long")

}
