% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{population}
\alias{population}
\title{World Population}
\format{A data.frame object with 38,654 rows and 3 columns
\describe{
  \item{\code{lat}, \code{lon}}{: Coordinates.}
  \item{\code{country_code}}{: Population.}
}}
\source{
\url{https://ecomfe.github.io/echarts-examples/public/data-gl/asset/data/population.json}
}
\usage{
population
}
\description{
A datatset of world population by tiles.
}
\keyword{datasets}
