% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scale_labels_size}
\alias{scale_labels_size}
\title{Scale Labels SIze}
\usage{
scale_labels_size(globe, min = 0.1, max = 1)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{min, max}{Target minimum and maximum values.}
}
\description{
Rescale \code{\link{globe_labels}} text size.
}
