% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scaling_altitude}
\alias{scaling_altitude}
\alias{scale_points_altitude}
\title{Scale Altitude}
\usage{
scale_points_altitude(globe, min = 0, max = 1)
}
\arguments{
\item{min, max}{Target minimum and maximum values of altitude.}
}
\description{
Rescale altitude to a more appropriate range, where 0 if flat
and 1 is the globe radius.
}
