% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scaling_dash}
\alias{scaling_dash}
\alias{scale_arc_dash_length}
\title{Scale Arc Dash}
\usage{
scale_arc_dash_length(globe, min = 0.1, max = 1)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{min, max}{Target minimum and maximum values of pixels.}
}
\description{
Rescale various aspects of arcs dashes.
}
\section{Scales}{

\itemize{
  \item{Dash length: where \code{1} is the entire length of the arc.}
}
}

\examples{
# basic
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_arcs(
    data = usflights, 
    coords (
      start_lat, start_lon, end_lat, end_lon,
      dash_length = cnt
    )
  ) \%>\% 
  scale_arc_dash_length()

}
