% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scaling_stroke}
\alias{scaling_stroke}
\alias{scale_arc_stroke}
\title{Scale Arc Stroke}
\usage{
scale_arc_stroke(globe, min = 0.1, max = 1)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{min, max}{Target minimum and maximum values of pixels.}
}
\description{
Rescale arc stroke to a more appropriate pixel values.
}
\examples{
# basic
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_arcs(
    data = usflights, 
    coords (
      start_lat, start_lon, end_lat, end_lon,
      stroke = cnt
    )
  ) \%>\% 
  scale_arc_stroke()

}
