% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{shared_background}
\alias{shared_background}
\title{Shared options}
\usage{
shared_background(value)
}
\arguments{
\item{value}{Value to set globally.}
}
\description{
Options that will be shared with all subsequent visualisation.
This is ideal when using Rmarkdown or Shiny, e.g.: only set
the background color once, instead of multiple times.
}
