% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{usflights}
\alias{usflights}
\title{U.S. Flights}
\format{A data.frame object with 178 rows and 8 columns
\describe{
  \item{\code{start_lat}, \code{start_lon}, \code{end_lat}, \code{end_lon}}{: geographical coordinates.}
  \item{\code{airline}}{; name of airline}
  \item{\code{airport_start}, \code{airport_end}}{: names of airports.}
  \item{\code{cnt}}{: number of flights}
}}
\usage{
usflights
}
\description{
A small data.frame of flight routes in the U.S.
}
\keyword{datasets}
