% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcs.R
\name{arcs_data}
\alias{arcs_data}
\alias{arcs_label}
\alias{arcs_start_lat}
\alias{arcs_start_lon}
\title{Arcs Data}
\usage{
arcs_data(globe, data)

arcs_label(globe, label)

arcs_start_lat(globe, lat)

arcs_start_lon(globe, lon)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame containing arcs data.}
}
\description{
Add arcs data to a globe.
}
