% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arcs.R
\name{arcs_data}
\alias{arcs_data}
\alias{arcs_label}
\alias{arcs_start_lat}
\alias{arcs_start_lon}
\alias{arcs_end_lat}
\alias{arcs_end_lon}
\alias{arcs_color}
\alias{arcs_altitude}
\alias{arcs_altitude_scale}
\alias{arcs_stroke}
\alias{arcs_curve_resolution}
\alias{arcs_circular_resolution}
\alias{arcs_dash_length}
\alias{arcs_dash_gap}
\alias{arcs_dash_initial_gap}
\alias{arcs_dash_animate}
\alias{arcs_transition}
\title{Arcs Data}
\usage{
arcs_data(globe, data)

arcs_label(globe, label)

arcs_start_lat(globe, lat = "startLat")

arcs_start_lon(globe, lon = "startLng")

arcs_end_lat(globe, lat = "endLat")

arcs_end_lon(globe, lon = "endLng")

arcs_color(globe, color = constant("#ffffaa"))

arcs_altitude(globe, altitude)

arcs_altitude_scale(globe, scale = 0.5)

arcs_stroke(globe, stroke)

arcs_curve_resolution(globe, curve = 64L)

arcs_circular_resolution(globe, circular = 6L)

arcs_dash_length(globe, length = 1L)

arcs_dash_gap(globe, gap = 0L)

arcs_dash_initial_gap(globe, gap = 0L)

arcs_dash_animate(globe, ms = 0L)

arcs_transition(globe, ms = 0L)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame containing arcs data.}

\item{label}{Column name containing label or a charcater vector. Supports html.}

\item{lat, lon}{Column names or numeric value indicating 
starting and ending coordinates of arc(s).}

\item{color}{Column name or character vector indicating color of points.}

\item{altitude}{Column name or a numeric constant for the arc's maximum altitude
(ocurring at the half-way distance between the two points) in terms of globe radius
units (0 = a numeric constant for the scale of the arc's automatic altitude, in terms of units of the great-arc distance between the two points. A value of 1 indicates the arc should be as high as its length on the ground. Only applicable if arcAltitude is not set.0 altitude (ground line), 1 = globe radius). If a value of null or
undefined is used, the altitude is automatically set proportionally to the distance
between the two points, according to the scale set in \code{arcs_altitude_scale}.}

\item{scale}{Column name or a numeric constant for the scale of the arc's automatic
altitude, in terms of units of the great-arc distance between the two points. A 
value of 1 indicates the arc should be as high as its length on the ground. 
Only applicable if \code{arcs_altitude} is not set.}

\item{stroke}{Column name indicating the line's diameter, 
in angular degrees. A value of null or undefined will render a 
\href{https://threejs.org/docs/#api/objects/Line}{ThreeJS} Line 
whose width is constant (1px) regardless of the camera distance. 
Otherwise, a 
\href{https://threejs.org/docs/#api/en/geometries/TubeGeometry}{TubeGeometry}
is used.}

\item{curve}{Resolution, expressed in how many straight line 
segments to divide the curve by. Higher values yield smoother curves.}

\item{circular}{Radial geometric resolution of each line, 
expressed in how many slice segments to divide the tube's circumference. 
Only applicable when using Tube geometries (defined \code{stroke}).}

\item{length}{Column name or numeric giving the length of the dashed 
segments in the arc, in terms of relative length of the whole line 
(\code{1} = full line length).}

\item{gap}{Column name or numeric giving the length of the gap between 
dash segments, in terms of relative line length.}

\item{ms}{Colum name or numeric for the time duration (in ms) to animate 
the motion of dash positions from the start to the end point for a full line 
length. A value of 0 disables the animation.}
}
\description{
Add arcs data to a globe.
}
\examples{
create_globe() \%>\% 
  globe_img_url() \%>\% 
  arcs_data(usflights) \%>\% 
  arcs_start_lat("start_lat") \%>\% 
  arcs_start_lon("start_lon") \%>\% 
  arcs_end_lat("end_lat") \%>\% 
  arcs_end_lon("end_lon")

}
