% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\title{Coordinates}
\usage{
coords(lat, lon, ...)
}
\arguments{
\item{lat, lon, ...}{List of name value pairs giving coordinates to map to
variables. The names for lat and lon coordinates are typically omitted because
they are so common; all other aspects must be named.}
}
\description{
Coordinates to use to draw layers on globe.
}
\section{Coordinates}{

Valid coordinates (depending on layer).
\itemize{
  \item{\code{lat}, \code{lon}}
  \item{\code{start_lat}, \code{start_lon}}
  \item{\code{end_lat}, \code{end_lon}}
  \item{\code{altitude}}
  \item{\code{radius}}
  \item{\code{color}}
  \item{\code{label}}
  \item{\code{resolution}}
  \item{\code{merge}}
  \item{\code{transition}}
  \item{\code{size}}
  \item{\code{text}}
  \item{\code{rotation}}
  \item{\code{include_dot}}
  \item{\code{dot_radius}}
  \item{\code{dot_orientation}}
  \item{\code{type_face}}
  \item{\code{altitude_scale}}
  \item{\code{stroke}}
  \item{\code{curve_resolution}}
  \item{\code{circular_resolution}}
  \item{\code{dash_length}}
  \item{\code{dash_gap}}
  \item{\code{dash_initial_gap}}
  \item{\code{dash_animate_time}}
  \item{\code{country}},
  \item{\code{side_color}, \code{cap_color}}
}
}

