% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.R
\name{coords}
\alias{coords}
\title{Coordinates}
\usage{
coords(lat, lon, ...)
}
\arguments{
\item{lat, lon, ...}{List of name value pairs giving coordinates to map to
variables. The names for lat and lon coordinates are typically omitted because
they are so common; all other aspects must be named.}
}
\description{
Coordinates to use to draw layers on globe.
}
\section{Coordinates}{

Valid coordinates (depending on layer).
\itemize{
  \item{\code{lat}, \code{lon}},
  \item{\code{altitude}}
  \item{\code{radius}}
  \item{\code{color}}
  \item{\code{label}}
  \item{\code{resolution}}
  \item{\code{merge}}
  \item{\code{transition}}
  \item{\code{size}}
  \item{\code{text}}
  \item{\code{rotation}}
  \item{\code{include_dot}}
  \item{\code{dot_radius}}
  \item{\code{dot_orientation}}
  \item{\code{type_face}}
}
}

