% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{globe_choropleth}
\alias{globe_choropleth}
\title{Choropleth}
\usage{
globe_choropleth(globe, ..., data = NULL, inherit_coords = TRUE,
  on_click = NULL, on_right_click = NULL, on_hover = NULL,
  match = c("auto", "iso2", "iso3", "name"))
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{...}{Coordinates, as specified by \code{\link{coords}}.}

\item{data}{A data.frame of points to draw.}

\item{inherit_coords}{Whether to inherit the coordinates (\code{\link{coords}})
from \code{\link{create_globe}}. Only applies to method applied to object of class
\code{globe4r} not on objects of class \code{globeProxy}.}

\item{on_click}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{on_right_click}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{on_hover}{JavaScript functions as \link[htmlwidgets]{JS} 
or \code{TRUE} to pick up the event from Shiny server side, see example.}

\item{match}{The type of \code{country} identifier, 
\code{auto} attempts to infer the type, \code{iso2}
(e.g.: "US"), iso3 (e.g.: "USA"), or the country name.}
}
\description{
Add points to a globe.
}
\section{Coordinates}{

Valid coordinates.
\itemize{
  \item{\code{country}},
  \item{\code{altitude}}
  \item{\code{label}}
  \item{\code{side_color}, \code{cap_color}}
  \item{\code{transition}}
}
}

\examples{
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_choropleth(
    data = agriland, 
    coords(country = country_code, cap_color = percent)
  )

}
