% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{globe_choropleth}
\alias{globe_choropleth}
\title{Choropleth}
\usage{
globe_choropleth(globe, data, country, cap_color = NULL,
  side_color = NULL, altitude = NULL, label = NULL,
  transition = 1000L, on_click = NULL, on_right_click = NULL,
  on_hover = NULL, match = c("auto", "iso2", "iso3", "name"))
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame of points to draw.}

\item{country}{The Bare column name containing either 
the ISO2, ISO3, or country name (see \code{match} argument).}

\item{cap_color}{Bare column name containing the color of 
the surface.}

\item{side_color}{Bare column name containing color of 
the sides.}

\item{altitude}{Altitude of countries where (0 = 0 altitude 
(flat polygon), 1 = globe radius).}

\item{label}{Bare column name containing label.}

\item{transition}{A value of 0 will size the cone immediatel
to their final altitude. New polygons are animated by rising
them from the ground up.}

\item{on_click, on_right_click, on_hover}{JavaScript functions as strings.}

\item{match}{The type of \code{country} identifier, 
\code{auto} attempts to infer the type, \code{iso2}
(e.g.: "US"), iso3 (e.g.: "USA"), or the country name.}
}
\description{
Add points to a globe.
}
\examples{
create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_choropleth(agriland, country_code, cap_color = percent)

}
