% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{globe_labels}
\alias{globe_labels}
\title{Labels}
\usage{
globe_labels(globe, ...)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{...}{Coordinates, as specified by \code{\link{coords}}.}
}
\description{
Add labels to a globe.
}
\section{Coordinates}{

Valid coordinates.
\itemize{
  \item{\code{lat}, \code{lon}},
  \item{\code{altitude}}
  \item{\code{color}}
  \item{\code{label}}
  \item{\code{resolution}}
  \item{\code{transition}}
  \item{\code{size}}
  \item{\code{text}}
  \item{\code{rotation}}
  \item{\code{include_dot}}
  \item{\code{dot_radius}}
  \item{\code{dot_orientation}}
  \item{\code{type_face}}
}
}

\examples{
quakes \%>\% 
  create_globe() \%>\% 
  globe_img_url() \%>\% 
  globe_labels(
    coords(lat, long, text = stations)
  )

library(shiny)

ui <- fluidPage(
  actionButton("add", "add quakes labels"),
  globeOutput("globe")
)

server <- function(input, output){
  output$globe <- renderGlobe({
    create_globe() \%>\% 
      globe_img_url() 
  })

  observeEvent(input$add, {
    globeProxy("globe") \%>\% 
      globe_labels(
        coords(lat, long, text = stations),
        data = quakes
      )
  })
}

if(interactive()) shinyApp(ui, server)

}
