% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{labels_data}
\alias{labels_data}
\alias{labels_lat}
\alias{labels_lon}
\alias{labels_text}
\alias{labels_color}
\alias{labels_altitude}
\alias{labels_size}
\alias{labels_typeface}
\alias{labels_rotation}
\alias{labels_resolution}
\alias{labels_include_dot}
\alias{labels_dot_radius}
\alias{labels_dot_orientation}
\alias{labels_transition}
\alias{labels_on_click}
\alias{labels_on_right_click}
\alias{labels_on_hover}
\title{Points Functional API}
\usage{
labels_data(globe, data)

labels_lat(globe, lat = "lat")

labels_lon(globe, lon = "lng")

labels_text(globe, text = "text")

labels_color(globe, color = constant("#ffffaa"))

labels_altitude(globe, altitude = 0L)

labels_size(globe, size = 0.5)

labels_typeface(globe, typeface)

labels_rotation(globe, rotation = 0L)

labels_resolution(globe, resolution = 3L)

labels_include_dot(globe, include = FALSE)

labels_dot_radius(globe, radius = 0.1)

labels_dot_orientation(globe, orientation = constant("bottom"))

labels_transition(globe, transition = 1000L)

labels_on_click(globe, func)

labels_on_right_click(globe, func)

labels_on_hover(globe, func)
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame of points to draw.}

\item{lat}{Column names or numeric value indicating coordinates.}

\item{lon}{Column names or numeric value indicating coordinates.}

\item{text}{Column name or constant of text.}

\item{color}{Column name or character vector indicating color of points.}

\item{altitude}{Column name or character vector indicating altitude of points 
in terms of globe radius units (0 = 0 altitude (flat circle), 1 = globe radius).}

\item{size}{Column name or constant of the size of the label text height, in angular degrees.}

\item{typeface}{Font of \code{text}. Supports any typeface font generated by 
\href{http://gero3.github.io/facetype.js/}{Facetype.js}.}

\item{rotation}{Column name or constant of the label rotation in degrees. 
The rotation is performed clockwise along the axis of its latitude parallel plane.}

\item{resolution}{Numeric value defining the geometric resolution of each 
cylinder, expressed in how many slice segments to divide the circumference. 
Higher values yield smoother cylinders.}

\item{include}{Boolean or colum name indicating whether to show the dot.}

\item{radius}{Column name of radius a numeric constant for the cylinder's 
radius, in angular degrees.}

\item{orientation}{Orientation of label, \code{right}, \code{top}, or \code{bottom}.}

\item{transition}{Duration (ms) of the transition 
to animate point changes involving geometry modifications. A value of 
\code{0} will move the objects immediately to their final position. 
New objects are animated by scaling them from the ground up. Only works 
if \code{merge} is disabled.}

\item{func}{JavaScript function as character vector.}
}
\description{
Functional API to add and customise points on globe.
}
\examples{
# use data
create_globe() \%>\% 
  labels_data(quakes) \%>\% 
  labels_lat("lat") \%>\% 
  labels_lon("long")

}
