% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{polygons_data}
\alias{polygons_data}
\alias{polygons_label}
\title{Polygons Data}
\usage{
polygons_data(globe, data)

polygons_label(globe, label = "name")
}
\arguments{
\item{globe}{An object of class \code{globe} as returned by
\code{\link{create_globe}}, or an object of class \code{globeProxy}
as returned by \code{\link{globeProxy}}.}

\item{data}{A data.frame containing arcs data.}

\item{label}{Column name containing label.}
}
\description{
Add polygons data to a globe.
}
